phase2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Todo List API Backend - Phase II"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.2",
]
